fix: rename check-changelog from .mjs to .ts
